move load group definitions to external classes
